[components] Close issue https://github.com/mozilla-mobile/android-components/issues/4876: Add UI for addon discovery and installation.

+++ mobile/android/android-components/components/feature/addons/src/main/java/mozilla/components/feature/addons/AddOn.kt
@@ -63,7 +63,7 @@ data class AddOn(
      * @property reviews The number of users that has scored this add-on.
      */
     data class Rating(
-        val average: Double,
+        val average: Float,
         val reviews: Int
     )
 }
